StorylineWebApp/requirements.txt
aiofiles==0.8.0
anyio==3.4.0
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
crispy-bootstrap5==0.6
Django==4.0.5
django-browser-reload==1.6.0
django-crispy-forms==1.14.0
fastapi==0.70.1
firebase==3.0.1
fontawesomefree==6.1.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
imutils==0.5.4
mysql-connector-python==8.0.27
mysql-connector-python-rf==2.2.2
numpy==1.21.4
opencv-python-headless==4.5.4.60
Pillow==9.1.1
protobuf==3.19.1
pycodestyle==2.8.0
pydantic==1.8.2
pymssql==2.2.3
PyMySQL==1.0.2
pyodbc==4.0.32
python-multipart==0.0.5
requests==2.26.0
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
sqlparse==0.4.2
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
